+++ source/overlay.ts
@@ -1,5 +1,5 @@
 import { createTextField } from "./elements/textField";
-import { HTMLView } from "./view/htmlView";
+import { HTMLView } from "./views/htmlView";
 
 /** The full-screen HTML overlay. */
 export default class Overlay {
