+++ source/views/schoolView.ts
